djangobackend/pyproject.toml
[tool.poetry]
name = "heic-wallpaper"
version = "0.1.0"
description = ""
authors = ["Thies <git@thies.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.4"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"
cryptography = "^44.0.0"
requests = "^2.32.3"
djangorestframework = "^3.15.2"
boto3 = "^1.35.78"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
